refactor: MAT-361 선수 상세 패널에서 Table, Profile 컴포넌트 추출
